Added beginings of integration tests.
So this adds a few tests using the child_process.fork() method,
in the attempt to recreate starting/ restarting conditions that
may occur for a service or application.

These have been broken into each job assignment method.

.each()
  - Ensures jobs do not get re-ran after restart.
  - Ensures array names still have jobs ran.

.schedule()
  - Ensures on startup future jobs do not get ran.
  - Ensures jobs scheduled before current time get ran.
  - Ensures array names still have jobs ran.

.every()
  - Ensures job gets ran as soon as possible.

There are a few nitpicks here that I still have, such as using
path.join() to ensure this can be tested in non-Unix enviroments.

Seeing as this uses Make, I am not 100% sure how useful
cross-compatibility is, not to mention I haven't validated that
the rest of the codebase uses path.join() either.