Fix: make `touch` promise-based (#667)
The [`job.touch()`](https://github.com/agenda/agenda/blob/ff94c8a4c9bc564a0bed9eaa79de1c4fdbed0fde/lib/job/touch.js#L10-L13) method didn't get migrated over to promises along with everything else and still expects a callback. However, when it passes that callback down to `job.save()`, an error is thrown because a callback is not allowed.

This managed to slip through because the test for it changed at some point to mock out the implementation of `job.save()` with a callback version that didn't throw when it was called with a callback.

https://github.com/agenda/agenda/blob/bd8a8e003cd09d6e9826accbf6c30be75212a9a9/test/job.js#L352-L364

This PR makes the behavior of `touch()` align with the rest of the library and updates the associated test to remove the mock on `save()`.