fix: _unlockJobs should retun when it resolves early
Calling `resolve` does not stop the function execution. This can lead to
the findAndMofidy call happening even though the promise has resolved
and the promise is no longer awaiting.

For instance in this code:

```
await agenda.stop()
await agenda.close()
```

We may find that we already closed the connection when the stop code is
still trying to reach the Mongo DB