fix: check if job has expired before we run it
it can happen that the process is so long in the queue, that it gets expired in the meantime.
if it is expired, it is put again in the queue, and then it could happen that a job that should
run with a concurrency of e.g. 1 runs in parallel twice.