Fix: Ensure cleared `nextRunAt` when saving to db
When using a mongodb client configured with the option `ignoreUndefined`
set to `true`, trying to clear the property `nextRunAt` from a job
scheduled using `schedule` method by setting it to `undefined` does not
take effect as it is ignored.

After computing `nextRunAt` property before running the job (`run.ts`), the
change in value is not persisted, making it always behind `lastRunAt`
after the first executing. This makes the program run the processor
very quickly and indefinitely, exhausting the CPU's capacity.

Setting the property to `null` at the beginning of the computation makes
sure the value is properly cleared.